feat: change optimization
